Add progress bar and a progress label,fixed bug where actual progress compares to expected and does not return the string "You are behind!"

+++ src/edu/bsu/sked/model/Assignment.java
@@ -185,7 +185,7 @@ public class Assignment {
 	public double getDaysPassedPercent() {
 		long numberOfAssignmentDays = Period.between(getStartDate(), getDueDate()).get(ChronoUnit.DAYS);
 		long numberOfDaysSinceStart = Period.between(getStartDate(), LocalDate.now()).get(ChronoUnit.DAYS);
-		return numberOfDaysSinceStart / numberOfAssignmentDays;
+		return (double)numberOfDaysSinceStart / numberOfAssignmentDays;
 	}
 
 	public void setSubtasks(List<Subtask> list) {
